fix: satisfy CI test and ESLint budget for parameterized runs
- Align WidgetTable confirm preview with template-only run hook params
- Extract template autocomplete, run form fields, and list drag hook
- Reduce ListFieldRenderer complexity to stay within ESLint budget
