+++ web_src/src/pages/workflowv2/mappers/start/runModal.tsx
@@ -1,12 +1,10 @@
 import React from "react";
 import Editor from "@monaco-editor/react";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { LoadingButton } from "@/components/ui/loading-button";
 import { showErrorToast } from "@/lib/toast";
-import { Checkbox } from "@/ui/checkbox";
 
+import { StartRunParameterFields } from "./startRunParameterFields";
 import {
   coerceParameterValue,
   initialParameterValue,
@@ -84,49 +82,11 @@ export function StartRunModal({
   return (
     <div className="space-y-4">
       {useParameterForm ? (
-        <div className="space-y-3">
-          {(parameters ?? []).map((param) => {
-            if (!param.name || !param.type) return null;
-            const id = `start-run-param-${param.name}`;
-            const label = parameterDisplayLabel(param);
-            return (
-              <div key={param.name} className="space-y-1.5">
-                {param.type === "boolean" ? (
-                  <div className="flex items-center gap-2">
-                    <Checkbox
-                      id={id}
-                      checked={Boolean(parameterValues[param.name])}
-                      onCheckedChange={(checked) =>
-                        setParameterValues((prev) => ({
-                          ...prev,
-                          [param.name]: checked === true,
-                        }))
-                      }
-                    />
-                    <Label htmlFor={id} className="cursor-pointer">
-                      {label}
-                    </Label>
-                  </div>
-                ) : (
-                  <>
-                    <Label htmlFor={id}>{label}</Label>
-                    <Input
-                      id={id}
-                      type={param.type === "number" ? "number" : "text"}
-                      value={String(parameterValues[param.name] ?? "")}
-                      onChange={(e) =>
-                        setParameterValues((prev) => ({
-                          ...prev,
-                          [param.name]: e.target.value,
-                        }))
-                      }
-                    />
-                  </>
-                )}
-              </div>
-            );
-          })}
-        </div>
+        <StartRunParameterFields
+          parameters={parameters ?? []}
+          parameterValues={parameterValues}
+          onParameterValuesChange={setParameterValues}
+        />
       ) : (
         <div className="border border-gray-200 dark:border-gray-700 rounded-md overflow-hidden">
           <Editor
